[Project64] Update file class

+++ Source/Common/FileClass.h
@@ -5,7 +5,7 @@
 class CFileBase
 {
 public:
-    enum OpenFlags 
+    enum OpenFlags
     {
         modeRead =          0x0000,
         modeWrite =         0x0001,
@@ -20,7 +20,7 @@ public:
         modeNoTruncate =    0x2000,
     };
 
-    enum Attribute 
+    enum Attribute
     {
         normal =    0x00,
         readOnly =  0x01,
@@ -36,7 +36,7 @@ public:
     virtual bool Open(const char * lpszFileName, uint32_t nOpenFlags ) = 0;
 
     virtual uint32_t GetPosition() const = 0;
-    virtual long Seek(long lOff, SeekPosition nFrom) = 0;
+    virtual int32_t Seek(int32_t lOff, SeekPosition nFrom) = 0;
     virtual bool SetLength(uint32_t dwNewLen) = 0;
     virtual uint32_t GetLength() const = 0;
 
@@ -74,7 +74,7 @@ public:
 
     // Overridables
     virtual uint32_t GetPosition() const;
-    virtual long Seek(long lOff, SeekPosition nFrom);
+    virtual int32_t Seek(int32_t lOff, SeekPosition nFrom);
     virtual bool SetLength(uint32_t dwNewLen);
     virtual uint32_t GetLength() const;
 
@@ -87,6 +87,6 @@ public:
     virtual bool SetEndOfFile();
 
 private:
-	CFile(const CFile&);				// Disable copy constructor
-	CFile& operator=(const CFile&);		// Disable assignment
+    CFile(const CFile&);				// Disable copy constructor
+    CFile& operator=(const CFile&);		// Disable assignment
 };
